requirements.txt
# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
mysql-connector-python==8.1.0

# Data Processing
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.1
xgboost==2.0.0
joblib==1.3.2

# Reporting
jinja2==3.1.2
weasyprint==60.1
matplotlib==3.8.0
seaborn==0.13.0

# Testing
pytest==7.4.2

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1 